+++ client/src/pages/auth/Register.jsx
@@ -1,16 +1,16 @@
 import React, { Fragment } from "react";
-import LandingLayout from "../../components/layout/LandingLayout";
 import { useLocation } from "react-router-dom";
+import LandingLayout from "../../old-ui-src/components/layout/LandingLayout";
 
-const LoginPage = () => {
+const RegisterPage = () => {
   const location = useLocation()
   return (
     <Fragment>
       <LandingLayout url={location.pathname}>
-        <div className="container">LOGIN PAGE</div>
+        <div>REGISTER PAGE</div>
       </LandingLayout>
     </Fragment>
   );
 };
 
-export default LoginPage;
+export default RegisterPage;
